+++ subprojects/library/src/commonMain/kotlin/org/kotools/types/EmailAddress.kt
@@ -161,9 +161,6 @@ public class EmailAddress private constructor(private val text: String) {
          * of the specified [text], or throws an [IllegalArgumentException] if
          * the string representation of [text] doesn't match the
          * [default pattern][PATTERN].
-         *
-         * See the [fromStringOrNull] method for returning `null` instead of
-         * throwing an exception in case of invalid [text].
          */
         @Deprecated(
             "Use the 'orThrow' method instead.",
@@ -185,9 +182,6 @@ public class EmailAddress private constructor(private val text: String) {
          * [text] doesn't match the string representation of the specified
          * [pattern], or if the string representation of [pattern] doesn't match
          * the [default pattern][PATTERN].
-         *
-         * See the [fromStringOrNull] method for returning `null` instead of
-         * throwing an exception in case of invalid [text] or [pattern].
          */
         @Deprecated(
             "Use the 'orThrow' method instead.",
@@ -214,37 +208,6 @@ public class EmailAddress private constructor(private val text: String) {
             return requireNotNull(address) {
                 InvalidEmailAddress("$text", pattern)
             }
-        }
-
-        /**
-         * Creates an instance of [EmailAddress] from the string representation
-         * of the specified [text].
-         * Returns `null` if the string representation of [text] doesn't match
-         * the string representation of the specified [pattern], or if the
-         * string representation of [pattern] doesn't match the
-         * [default pattern][PATTERN].
-         *
-         * This method is not available from Java code due to its non-explicit
-         * [support for nullable types](https://kotlinlang.org/docs/java-to-kotlin-nullability-guide.html#support-for-nullable-types).
-         *
-         * See the [fromString] method for throwing an exception instead of
-         * returning `null` in case of invalid [text] or [pattern].
-         */
-        @Deprecated(
-            "Use the 'orNull' method instead.",
-            ReplaceWith(
-                "EmailAddress.orNull(\"\$text\", \"\$pattern\")",
-                "org.kotools.types.EmailAddress"
-            ),
-            DeprecationLevel.ERROR
-        )
-        @DeprecatedAsErrorSince(KotoolsTypesVersion.V4_5_3)
-        @JvmSynthetic
-        public fun fromStringOrNull(text: Any, pattern: Any): EmailAddress? {
-            val patternAsString: String = pattern.toString()
-            val defaultRegex = Regex(PATTERN)
-            if (!patternAsString.matches(defaultRegex)) return null
-            return fromStringOrNull(text, patternAsString)
         }
 
         private fun fromStringOrNull(